+++ tests/model/test_store_cardreader_result.py
@@ -160,6 +160,36 @@ def entry_2(db, event_id, class_id):
 
 
 @pytest.fixture
+def entry_2_with_result(db, event_id, class_id, entry_2):
+    db.update_entry_result(
+        id=entry_2["id"],
+        chip=entry_2["chip"],
+        start_time=entry_2["start"].start_time,
+        result=PersonRaceResult(
+            status=ResultStatus.OK,
+            start_time=s1,
+            finish_time=f1,
+            punched_start_time=s1,
+            punched_finish_time=f1,
+            time=t(s1, f1),
+            split_times=[
+                SplitTime(
+                    control_code="101", punch_time=c1, time=t(s1, c1), status="OK"
+                ),
+                SplitTime(
+                    control_code="102", punch_time=c2, time=t(s1, c2), status="OK"
+                ),
+                SplitTime(
+                    control_code="103", punch_time=c3, time=t(s1, c3), status="OK"
+                ),
+            ],
+        ),
+    )
+    entry = db.get_entry(id=entry_2["id"])[0]
+    return copy.deepcopy(entry)
+
+
+@pytest.fixture
 def entry_3(db, event_id, class_id):
     id = db.add_entry(
         event_id=event_id,
@@ -181,45 +211,23 @@ def entry_3(db, event_id, class_id):
 
 
 @pytest.fixture
-def entry_4(db, event_id, class_id):
-    id = db.add_entry(
-        event_id=event_id,
-        competitor_id=None,
-        first_name="Yogi",
-        last_name="Löw",
-        gender="N",
-        year=None,
-        class_id=class_id,
-        club_id=None,
-        not_competing=False,
-        chip="7410",
-        fields={},
-        status=ResultStatus.INACTIVE,
-        start_time=None,
+def unassigned_entry(db, event_id):
+    result = PersonRaceResult(
+        status=ResultStatus.FINISHED,
+        punched_start_time=s1,
+        punched_finish_time=f1,
+        time=None,
+        split_times=[
+            SplitTime(control_code="101", punch_time=c1, status="Additional"),
+            SplitTime(control_code="102", punch_time=c2, status="Additional"),
+            SplitTime(control_code="103", punch_time=c3, status="Additional"),
+        ],
     )
-    db.update_entry_result(
-        id=id,
+    id = db.add_entry_result(
+        event_id=event_id,
         chip="7410",
+        result=result,
         start_time=None,
-        result=PersonRaceResult(
-            status=ResultStatus.OK,
-            start_time=s1,
-            finish_time=f1,
-            punched_start_time=s1,
-            punched_finish_time=f1,
-            time=t(s1, f1),
-            split_times=[
-                SplitTime(
-                    control_code="101", punch_time=c1, time=t(s1, c1), status="OK"
-                ),
-                SplitTime(
-                    control_code="102", punch_time=c2, time=t(s1, c2), status="OK"
-                ),
-                SplitTime(
-                    control_code="103", punch_time=c3, time=t(s1, c3), status="OK"
-                ),
-            ],
-        ),
     )
     entry = db.get_entry(id=id)[0]
     return copy.deepcopy(entry)
@@ -245,7 +253,7 @@ def assert_entries_are_equal(entry_a: Dict, entry_b: Dict) -> None:
     assert entry_a.get("missingControls", None) == entry_b.get("missingControls", None)
 
 
-def test_store_cardreader_result_if_cardnumber_is_unique(
+def test_assign_to_entry_if_cardnumber_is_unique(
     db, event_id, entry_1, entry_2, entry_3
 ):
     result = PersonRaceResult(
@@ -306,7 +314,7 @@ def test_store_cardreader_result_if_cardnumber_is_unique(
     assert_entries_are_equal(entry_a=entries[2], entry_b=entry_3)
 
 
-def test_store_cardreader_result_if_cardnumber_is_unique_but_finish_time_is_missing(
+def test_assign_to_entry_if_cardnumber_is_unique_but_finish_time_is_missing(
     db, event_id, entry_1, entry_2, entry_3
 ):
     result = PersonRaceResult(
@@ -367,7 +375,7 @@ def test_store_cardreader_result_if_cardnumber_is_unique_but_finish_time_is_miss
     assert_entries_are_equal(entry_a=entries[2], entry_b=entry_3)
 
 
-def test_store_cardreader_result_if_cardnumber_is_unique_but_start_time_is_missing(
+def test_assign_to_entry_if_cardnumber_is_unique_but_start_time_is_missing(
     db, event_id, entry_1, entry_2, entry_3
 ):
     result = PersonRaceResult(
@@ -428,7 +436,7 @@ def test_store_cardreader_result_if_cardnumber_is_unique_but_start_time_is_missi
     assert_entries_are_equal(entry_a=entries[2], entry_b=entry_3)
 
 
-def test_store_cardreader_result_if_cardnumber_is_unique_but_controls_are_missing(
+def test_assign_to_entry_if_cardnumber_is_unique_but_controls_are_missing(
     db, event_id, entry_1, entry_2, entry_3
 ):
     result = PersonRaceResult(
@@ -487,7 +495,202 @@ def test_store_cardreader_result_if_cardnumber_is_unique_but_controls_are_missin
     assert_entries_are_equal(entry_a=entries[2], entry_b=entry_3)
 
 
-def test_store_cardreader_result_if_cardnumber_is_unknown(
+def test_assign_to_entry_and_delete_unnamed_entry_with_same_result(
+    db, event_id, entry_1, entry_2, entry_3, unassigned_entry
+):
+    result = PersonRaceResult(
+        status=ResultStatus.FINISHED,
+        punched_start_time=s1,
+        punched_finish_time=f1,
+        time=None,
+        split_times=[
+            SplitTime(control_code="101", punch_time=c1, status="Additional"),
+            SplitTime(control_code="102", punch_time=c2, status="Additional"),
+            SplitTime(control_code="103", punch_time=c3, status="Additional"),
+        ],
+    )
+    item = {
+        "entryType": "cardRead",
+        "entryTime": entry_time,
+        "controlCard": "7410",
+        "result": result,
+    }
+
+    status, event, res = model.store_cardreader_result(event_key="4711", item=item)
+    print(res)
+
+    assert status == "cardRead"
+    assert event["id"] == event_id
+    assert res == {
+        "entryTime": entry_time,
+        "eventId": event_id,
+        "controlCard": "7410",
+        "firstName": "Yogi",
+        "lastName": "Löw",
+        "club": None,
+        "class": "Elite",
+        "status": ResultStatus.OK,
+        "time": t(s1, f1),
+        "error": None,
+        "missingControls": [],
+    }
+
+    entries = db.get_entries(event_id=event_id)
+    assert len(entries) == 3
+
+    entry_2["result"] = PersonRaceResult(
+        status=ResultStatus.OK,
+        start_time=s1,
+        finish_time=f1,
+        punched_start_time=s1,
+        punched_finish_time=f1,
+        time=t(s1, f1),
+        split_times=[
+            SplitTime(control_code="101", punch_time=c1, time=t(s1, c1), status="OK"),
+            SplitTime(control_code="102", punch_time=c2, time=t(s1, c2), status="OK"),
+            SplitTime(control_code="103", punch_time=c3, time=t(s1, c3), status="OK"),
+        ],
+    )
+
+    assert_entries_are_equal(entry_a=entries[0], entry_b=entry_1)
+    assert_entries_are_equal(entry_a=entries[1], entry_b=entry_2)
+    assert_entries_are_equal(entry_a=entries[2], entry_b=entry_3)
+
+
+def test_store_as_new_entry_if_another_result_exists(
+    db, event_id, entry_1, entry_2, entry_3, unassigned_entry
+):
+    result = PersonRaceResult(
+        status=ResultStatus.FINISHED,
+        punched_start_time=s1,
+        punched_finish_time=f1,
+        time=None,
+        split_times=[
+            SplitTime(control_code="101", punch_time=c1, status="Additional"),
+            SplitTime(control_code="103", punch_time=c3, status="Additional"),
+        ],
+    )
+    item = {
+        "entryType": "cardRead",
+        "entryTime": entry_time,
+        "controlCard": "7410",
+        "result": result,
+    }
+
+    status, event, res = model.store_cardreader_result(event_key="4711", item=item)
+
+    assert status == "cardRead"
+    assert event["id"] == event_id
+    assert res == {
+        "entryTime": entry_time,
+        "eventId": event_id,
+        "controlCard": "7410",
+        "firstName": None,
+        "lastName": None,
+        "club": None,
+        "class": None,
+        "status": ResultStatus.FINISHED,
+        "time": None,
+        "error": "There are other results for this card",
+    }
+
+    entries = db.get_entries(event_id=event_id)
+    assert len(entries) == 5
+
+    result = PersonRaceResult(
+        status=ResultStatus.FINISHED,
+        start_time=s1,
+        finish_time=f1,
+        punched_start_time=s1,
+        punched_finish_time=f1,
+        time=t(s1, f1),
+        split_times=[
+            SplitTime(
+                control_code="101", punch_time=c1, time=t(s1, c1), status="Additional"
+            ),
+            SplitTime(
+                control_code="103", punch_time=c3, time=t(s1, c3), status="Additional"
+            ),
+        ],
+    )
+    new_entry = {
+        "id": [
+            e["id"]
+            for e in entries
+            if e["id"]
+            not in [entry_1["id"], entry_2["id"], entry_3["id"], unassigned_entry["id"]]
+        ][0],
+        "event_id": event_id,
+        "first_name": None,
+        "last_name": None,
+        "gender": None,
+        "year": None,
+        "competitor_id": None,
+        "not_competing": False,
+        "chip": "7410",
+        "fields": {},
+        "result": result,
+        "start": PersonRaceStart(),
+        "club_id": None,
+        "club": None,
+        "class_id": None,
+        "class_": None,
+    }
+
+    assert_entries_are_equal(entry_a=entries[0], entry_b=unassigned_entry)
+    assert_entries_are_equal(entry_a=entries[1], entry_b=new_entry)
+    assert_entries_are_equal(entry_a=entries[2], entry_b=entry_1)
+    assert_entries_are_equal(entry_a=entries[3], entry_b=entry_2)
+    assert_entries_are_equal(entry_a=entries[4], entry_b=entry_3)
+
+
+def test_do_not_store_as_new_entry_if_result_already_exists(
+    db, event_id, entry_1, entry_3, unassigned_entry
+):
+    result = PersonRaceResult(
+        status=ResultStatus.FINISHED,
+        punched_start_time=s1,
+        punched_finish_time=f1,
+        time=None,
+        split_times=[
+            SplitTime(control_code="101", punch_time=c1, status="Additional"),
+            SplitTime(control_code="102", punch_time=c2, status="Additional"),
+            SplitTime(control_code="103", punch_time=c3, status="Additional"),
+        ],
+    )
+    item = {
+        "entryType": "cardRead",
+        "entryTime": entry_time,
+        "controlCard": "7410",
+        "result": result,
+    }
+
+    status, event, res = model.store_cardreader_result(event_key="4711", item=item)
+
+    assert status == "cardRead"
+    assert event["id"] == event_id
+    assert res == {
+        "entryTime": entry_time,
+        "eventId": event_id,
+        "controlCard": "7410",
+        "firstName": None,
+        "lastName": None,
+        "club": None,
+        "class": None,
+        "status": ResultStatus.FINISHED,
+        "time": None,
+        "error": "Control card unknown",
+    }
+
+    entries = db.get_entries(event_id=event_id)
+    assert len(entries) == 3
+
+    assert_entries_are_equal(entry_a=entries[0], entry_b=unassigned_entry)
+    assert_entries_are_equal(entry_a=entries[1], entry_b=entry_1)
+    assert_entries_are_equal(entry_a=entries[2], entry_b=entry_3)
+
+
+def test_store_as_new_entry_if_cardnumber_is_unknown(
     db, event_id, entry_1, entry_2, entry_3
 ):
     result = PersonRaceResult(
@@ -576,7 +779,7 @@ def test_store_cardreader_result_if_cardnumber_is_unknown(
     assert_entries_are_equal(entry_a=entries[3], entry_b=entry_3)
 
 
-def test_store_cardreader_result_if_cardnumber_exist_several_times(
+def test_store_as_new_entry_if_cardnumber_exist_several_times(
     db, event_id, entry_1, entry_2, entry_3
 ):
     result = PersonRaceResult(
@@ -611,7 +814,7 @@ def test_store_cardreader_result_if_cardnumber_exist_several_times(
         "class": None,
         "status": ResultStatus.FINISHED,
         "time": None,
-        "error": "Control card assigned several times",
+        "error": "There are several entries for this card",
     }
 
     entries = db.get_entries(event_id=event_id)
@@ -665,8 +868,8 @@ def test_store_cardreader_result_if_cardnumber_exist_several_times(
     assert_entries_are_equal(entry_a=entries[3], entry_b=entry_3)
 
 
-def test_store_cardreader_result_if_cardnumer_is_unique_with_same_result(
-    db, event_id, entry_1, entry_4, entry_3
+def test_use_already_assigned_entry_if_it_has_the_same_result(
+    db, event_id, entry_1, entry_2_with_result, entry_3
 ):
     result = PersonRaceResult(
         status=ResultStatus.FINISHED,
@@ -708,12 +911,12 @@ def test_store_cardreader_result_if_cardnumer_is_unique_with_same_result(
     assert len(entries) == 3
 
     assert_entries_are_equal(entry_a=entries[0], entry_b=entry_1)
-    assert_entries_are_equal(entry_a=entries[1], entry_b=entry_4)
+    assert_entries_are_equal(entry_a=entries[1], entry_b=entry_2_with_result)
     assert_entries_are_equal(entry_a=entries[2], entry_b=entry_3)
 
 
-def test_store_cardreader_result_if_cardnumber_is_unique_with_another_result(
-    db, event_id, entry_1, entry_4, entry_3
+def test_store_as_new_entry_if_cardnumber_is_unique_with_another_result(
+    db, event_id, entry_1, entry_2_with_result, entry_3
 ):
     result = PersonRaceResult(
         status=ResultStatus.FINISHED,
@@ -747,7 +950,7 @@ def test_store_cardreader_result_if_cardnumber_is_unique_with_another_result(
         "class": None,
         "status": ResultStatus.FINISHED,
         "time": None,
-        "error": "There is already a result",
+        "error": "There are other results for this card",
     }
 
     entries = db.get_entries(event_id=event_id)
@@ -776,7 +979,7 @@ def test_store_cardreader_result_if_cardnumber_is_unique_with_another_result(
         "id": [
             e["id"]
             for e in entries
-            if e["id"] not in [entry_1["id"], entry_4["id"], entry_3["id"]]
+            if e["id"] not in [entry_1["id"], entry_2_with_result["id"], entry_3["id"]]
         ][0],
         "event_id": event_id,
         "first_name": None,
@@ -797,11 +1000,11 @@ def test_store_cardreader_result_if_cardnumber_is_unique_with_another_result(
 
     assert_entries_are_equal(entry_a=entries[0], entry_b=new_entry)
     assert_entries_are_equal(entry_a=entries[1], entry_b=entry_1)
-    assert_entries_are_equal(entry_a=entries[2], entry_b=entry_4)
+    assert_entries_are_equal(entry_a=entries[2], entry_b=entry_2_with_result)
     assert_entries_are_equal(entry_a=entries[3], entry_b=entry_3)
 
 
-def test_store_cardreader_result_use_empty_control_list_if_course_is_undefined(
+def test_use_empty_control_list_if_course_is_undefined(
     db, event_id, entry_1_without_course
 ):
     result = PersonRaceResult(
@@ -866,7 +1069,7 @@ def test_store_cardreader_result_use_empty_control_list_if_course_is_undefined(
     assert_entries_are_equal(entry_a=entries[0], entry_b=entry_1_without_course)
 
 
-def test_store_cardreader_result_exception_if_eventkey_is_unknown(
+def test_raise_exception_if_event_key_is_unknown(
     db, event_id, entry_1, entry_2, entry_3
 ):
     result = PersonRaceResult(